requirements.txt
torch
torchvision
git+https://github.com/openai/CLIP.git
ultralytics
flask
faiss-cpu
opencv-python
Pillow
numpy
torchreid
matplotlib
